Исправлены замечания: Шаги по созданию, получению заказа перенесены в OrderClient; Поле исправлено на приватное для принципа инкапсуляции; В негативный тест по созданию юзера добавлено удаление юзера.
