+++ src/main/java/ru/yandex/praktikum/model/Order.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Order {
-    List <String> ingredients = new ArrayList<>();
+    private List <String> ingredients = new ArrayList<>();
 
     public Order(List<String> ingredients) {
         this.ingredients = ingredients;
